+++ src/pytest/test_rsa.py
@@ -17,7 +17,7 @@ def test_encrypt(rsa_key):
     assert rsa.encrypt("abcd",rsa_key["e"],rsa_key["n"]) == 'b8f2c1456e5d44b0762bdb489298c399'
 
 def test_decrypt(rsa_key):
-    assert rsa.decrypt("b8f2c1456e5d44b0762bdb489298c399", rsa_key["d"], rsa_key["n"])
+    assert rsa.decrypt("b8f2c1456e5d44b0762bdb489298c399", rsa_key["d"], rsa_key["n"]) == "abcd"
 
 def test_genKey():
     key = rsa.genKey(256)